solved the problem about IOAPIC mapping. It was not map because the address was truncated to 32bit.

+++ arch/x86/include/asm/io_apic.h
@@ -136,7 +136,7 @@ struct IR_IO_APIC_route_entry {
 extern int nr_ioapics;
 
 extern int mpc_ioapic_id(int ioapic);
-extern unsigned int mpc_ioapic_addr(int ioapic);
+extern unsigned  long mpc_ioapic_addr(int ioapic);
 extern struct mp_ioapic_gsi *mp_ioapic_gsi_routing(int ioapic);
 
 #define MP_MAX_IOAPIC_PIN 127
